refactor: Move vulnerability serialization code from copy to vulnerability
Serializing vulnerability functions seems to better fit closer to the vulnerability function
