+++ modules/lib/src/main/kotlin/dev/vulnlog/lib/core/Vulnerability.kt
@@ -4,6 +4,8 @@
 package dev.vulnlog.lib.core
 
 import dev.vulnlog.lib.model.VulnId
+import dev.vulnlog.lib.parse.v1.dto.VulnerabilityEntryDto
+import tools.jackson.databind.ObjectMapper
 
 /**
  * Parses a vulnerability ID string and returns the corresponding VulnId instance.
@@ -30,3 +32,31 @@ fun parseVulnId(id: String): VulnId =
  * @return The uppercase string representation of the VulnId, based on its specific type's identifier.
  */
 fun VulnId.canonical(): String = id.uppercase()
+
+/**
+ * Serializes a [VulnerabilityEntryDto] object to a YAML-compatible string representation
+ * with specific indentation adjustments for proper formatting.
+ *
+ * The YAML format adjusts the first line with a prefix `-` and indents subsequent lines with spaces.
+ * This ensures that the resulting YAML is appropriately nested for inclusion in larger YAML structures.
+ *
+ * @param entry The [VulnerabilityEntryDto] object to be serialized.
+ * @param mapper An instance of [ObjectMapper] used for converting the entry object into a raw JSON string.
+ * @return A string representing the YAML-compatible serialized format of the given entry.
+ */
+fun serializeEntryYaml(
+    entry: VulnerabilityEntryDto,
+    mapper: ObjectMapper,
+): String {
+    val raw = mapper.writeValueAsString(entry)
+    val lines =
+        raw
+            .lines()
+            .dropWhile { it.trimStart().startsWith("---") || it.isBlank() }
+            .dropLastWhile { it.isBlank() }
+
+    return lines
+        .mapIndexed { index, line ->
+            if (index == 0) "  - $line" else "    $line"
+        }.joinToString("\n")
+}
