+++ modules/lib/src/test/kotlin/dev/vulnlog/lib/core/VulnerabilityTest.kt
@@ -4,9 +4,14 @@
 package dev.vulnlog.lib.core
 
 import dev.vulnlog.lib.model.VulnId
+import dev.vulnlog.lib.parse.createYamlMapper
+import dev.vulnlog.lib.parse.v1.dto.ReportEntryDto
+import dev.vulnlog.lib.parse.v1.dto.VulnerabilityEntryDto
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.shouldBe
+import io.kotest.matchers.string.shouldContain
+import io.kotest.matchers.string.shouldNotContain
 import io.kotest.matchers.types.shouldBeInstanceOf
 
 class VulnerabilityTest :
@@ -83,6 +88,68 @@ class VulnerabilityTest :
 
             test("already uppercase CVE stays uppercase") {
                 VulnId.Cve("CVE-2021-44228").canonical() shouldBe "CVE-2021-44228"
+            }
+        }
+
+        context("serializeEntryYaml") {
+
+            test("produces correctly indented YAML list item") {
+                val dto =
+                    VulnerabilityEntryDto(
+                        id = "CVE-2026-1234",
+                        releases = listOf("1.0.0"),
+                        packages = listOf("pkg:npm/example-lib@2.3.0"),
+                        reports = listOf(ReportEntryDto(reporter = "trivy")),
+                        verdict = "not affected",
+                        justification = "vulnerable code not in execute path",
+                    )
+
+                val yaml = serializeEntryYaml(dto, createYamlMapper())
+
+                yaml shouldContain "  - id: "
+                yaml shouldContain "    releases:"
+                yaml shouldContain "    verdict: "
+                yaml shouldNotContain "---"
+            }
+
+            test("omits null fields") {
+                val dto =
+                    VulnerabilityEntryDto(
+                        id = "CVE-2026-1234",
+                        releases = listOf("1.0.0"),
+                        packages = listOf("pkg:npm/example-lib@2.3.0"),
+                        reports = listOf(ReportEntryDto(reporter = "trivy")),
+                    )
+
+                val yaml = serializeEntryYaml(dto, createYamlMapper())
+
+                yaml shouldNotContain "verdict:"
+                yaml shouldNotContain "severity:"
+                yaml shouldNotContain "justification:"
+                yaml shouldNotContain "analysis:"
+                yaml shouldNotContain "description:"
+                yaml shouldNotContain "comment:"
+                yaml shouldNotContain "resolution:"
+                yaml shouldNotContain "source:"
+                yaml shouldNotContain "suppress:"
+            }
+
+            test("omits empty lists") {
+                val dto =
+                    VulnerabilityEntryDto(
+                        id = "CVE-2026-1234",
+                        releases = listOf("1.0.0"),
+                        packages = listOf("pkg:npm/example-lib@2.3.0"),
+                        reports = listOf(ReportEntryDto(reporter = "trivy")),
+                        aliases = emptyList(),
+                        tags = emptyList(),
+                    )
+
+                val yaml = serializeEntryYaml(dto, createYamlMapper())
+
+                yaml shouldNotContain "aliases:"
+                yaml shouldNotContain "tags:"
+                yaml shouldNotContain "vuln_ids:"
             }
         }
     })
